+++ test/fortnox_ex/models/customer_test.exs
@@ -6,57 +6,57 @@ defmodule FortnoxEx.Models.CustomerTest do
   defp access_token, do: System.fetch_env!("ACCESS_TOKEN")
 
   @tag :skip
-  describe "create_currency/2" do
-    test "creates a currency" do
+  describe "create_customer/2" do
+    test "creates a customer" do
       client = FortnoxEx.client(client_secret(), access_token())
-      {:ok, currency} = FortnoxEx.create_currency(client, %{"Name" => "Test Corporation"})
-      assert currency["Name"] == "Test Corporation"
+      {:ok, customer} = FortnoxEx.create_customer(client, %{"Name" => "Test Corporation"})
+      assert customer["Name"] == "Test Corporation"
     end
   end
 
   @tag :skip
-  describe "list_currencies/2" do
-    test "fetches currencies" do
+  describe "list_customers/2" do
+    test "fetches customers" do
       client = FortnoxEx.client(client_secret(), access_token())
-      {:ok, meta, currencies} = FortnoxEx.list_currencies(client, [])
+      {:ok, meta, customers} = FortnoxEx.list_customers(client, [])
       assert meta == %{"@CurrentPage" => 1, "@TotalPages" => 0, "@TotalResources" => 0}
-      assert Enum.count(currencies) == 0
+      assert Enum.count(customers) == 0
     end
   end
 
   @tag :skip
-  describe "stream_currencies/2" do
-    test "streams currencies" do
+  describe "stream_customers/2" do
+    test "streams customers" do
       client = FortnoxEx.client(client_secret(), access_token())
-      stream = FortnoxEx.stream_currencies(client, [limit: 5])
+      stream = FortnoxEx.stream_customers(client, [limit: 5])
       [first | _] = Enum.to_list(stream)
       assert first["CustomerNumber"] == "1"
     end
   end
 
   @tag :skip
-  describe "get_currency/2" do
-    test "fetches a currency" do
+  describe "get_customer/2" do
+    test "fetches a customer" do
       client = FortnoxEx.client(client_secret(), access_token())
-      {:ok, currency} = FortnoxEx.get_currency(client, "1")
-      assert currency["CustomerNumber"] == "1"
+      {:ok, customer} = FortnoxEx.get_customer(client, "1")
+      assert customer["CustomerNumber"] == "1"
     end
   end
 
   @tag :skip
-  describe "update_currency/2" do
-    test "updates a currency" do
+  describe "update_customer/2" do
+    test "updates a customer" do
       client = FortnoxEx.client(client_secret(), access_token())
-      {:ok, currency} = FortnoxEx.update_currency(client, "1", %{"Name" => "Test Corporation"})
-      assert currency["Name"] == "Test Corporation"
+      {:ok, customer} = FortnoxEx.update_customer(client, "1", %{"Name" => "Test Corporation"})
+      assert customer["Name"] == "Test Corporation"
     end
   end
 
   @tag :skip
-  describe "delete_currency/2" do
-    test "deletes a currency" do
+  describe "delete_customer/2" do
+    test "deletes a customer" do
       client = FortnoxEx.client(client_secret(), access_token())
-      :ok = FortnoxEx.delete_currency(client, "1")
+      :ok = FortnoxEx.delete_customer(client, "1")
     end
   end
 end